pyproject.toml
[project]
name = "specify-cli"
version = "0.0.4"
description = "Setup tool for Specify spec-driven development projects"
requires-python = ">=3.11"
dependencies = [
    # Existing spec-kit dependencies
    "typer",
    "rich>=13.0.0",
    "httpx[socks]",
    "platformdirs",
    "readchar",
    "truststore>=0.10.4",
    # APM dependencies (from awd-cli, excluding runtime/embargo items)
    "click>=8.0.0",
    "colorama>=0.4.6",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "python-frontmatter>=1.0.0",
    "tomli>=1.2.0; python_version<'3.11'",
    "toml>=0.10.2",
    "rich-click>=1.7.0",
    "watchdog>=3.0.0",
    "GitPython>=3.1.0",
]

[project.scripts]
specify = "specify_cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specify_cli", "src/apm_cli"]
